Make infura setup work with https
I was trying to use infura-like setup with `ws` (Web sockets). Since
`ws` sends plain-text, it does not work in prouduction. We cannot use
`wss` since it is not supported natively by geth
 - https://github.com/ethereum/go-ethereum/issues/16423
Even K8s only supports https. To avoid this issue, I shifted to https.
Subscriptions don't work with https, so, a crude polling mechanism to
check for transaction confirmations.

Also, in this commit, change the nonce count to include "pending"
transactions, so that, multiple transactions can be sent without waiting
for completion.