Do not send reveal transactions in parallel
The reveal transactions, up to 3, were sent in parallel and in infura
setup, they will end up with the same nonce, so, only one of them will
be confirmed.

Rather than sending them in parallel, I added a one second and a two
seconds delay for the second and third transaction. This, in practice,
seems to ensure that these transactions get the correct nonce.