Handle nonce related to simultaneous transactions
Handle the case when too many transactions are sent close to each other.

Geth has `web3.eth.getTransactionCount(account, 'pending')` but even
that only seems to work once the transactions have made to the
transaction pool. Therefore, for transactions sent too close to each
other, which is the case for SMS verification transactions, maintain a
local nonce and ensure that new Nonce is equal or greater than local
nonce.